fix(onboarding): refactor
